backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
python-jose==3.3.0
python-multipart==0.0.7
pyjwt==2.8.0
